fix: fetch data to save on end session inside component instead of Header (#225)
